+++ hss_okta_flutter/lib/src/web/hss_okta_flutter_web.dart
@@ -82,7 +82,7 @@ class HssOktaFlutterWeb extends HssOktaFlutterWebPlatformInterface {
   ///[key]- Unique key to store the token in the tokenManager. This is used later when you want to get, delete, or renew the token.
   /// [token] - Token object that will be added
 
-  void addToken(String key, Token token) async {
+  void addToken(String key, AbstractToken token) async {
     await promiseToFuture(_auth.tokenManager.add(key, token));
   }
 
@@ -101,7 +101,7 @@ class HssOktaFlutterWeb extends HssOktaFlutterWebPlatformInterface {
   /// The [Token] object will be returned if it exists in storage. Tokens will be removed from storage if they have expired and autoRenew is false or if there was an error while renewing the token.
   /// The [TokenManager] will emit a removed event when tokens are removed.
 
-  Future<Token> getToken(String key) async {
+  Future<AbstractToken> getToken(String key) async {
     final res = await promiseToFuture(_auth.tokenManager.get(key));
     return res;
   }
@@ -137,6 +137,15 @@ class HssOktaFlutterWeb extends HssOktaFlutterWebPlatformInterface {
     return _auth.getIdToken();
   }
 
+  /// Retrieve the details about a user.
+  ///
+  /// [accessTokenObject] - (optional) an access token returned by this library.
+  ///  [idTokenObject] - (optional) an ID token.
+  ///
+  /// By default, if no parameters are passed, both the access token and ID token objects will be retrieved from the TokenManager.
+  ///  It is assumed that the access token is stored using the key "accessToken" and the ID token is stored under the key "idToken".
+  ///  If you have stored either token in a non-standard location, this logic can be skipped by passing the access and ID token objects directly.
+
   Future<Map> getUserInfo(
       {AccessToken? accessTokenObject, IDToken? idTokenObject}) async {
     final res = await promiseToFuture(_auth.token.getUserInfo(
@@ -151,4 +160,7 @@ class HssOktaFlutterWeb extends HssOktaFlutterWebPlatformInterface {
   Future<TokenResponse> renewToken({required String refreshToken}) {
     throw UnimplementedError();
   }
+
+  Future<bool> hasTokenExpired(AbstractToken token) async =>
+      _auth.tokenManager.hasExpired(token);
 }